requirements.txt
feedparser==6.0.8
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
PyQt5==5.15.4
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.23.0
setuptools~=41.2.0
Pillow~=8.4.0